+++ xs/include/xs/mir.h
@@ -65,6 +65,8 @@ typedef enum
   XS_MIR_INSTRUCTION_REM_I32,
   XS_MIR_INSTRUCTION_AND_I32,
   XS_MIR_INSTRUCTION_OR_I32,
+  XS_MIR_INSTRUCTION_SHL_I32,
+  XS_MIR_INSTRUCTION_SHR_I32,
   XS_MIR_INSTRUCTION_EQ_I32,
   XS_MIR_INSTRUCTION_NE_I32,
   XS_MIR_INSTRUCTION_LT_I32,
@@ -143,6 +145,10 @@ XsMirStatus xs_mir_block_and_i32(XsMirBlock *block, XsMirValueId left, XsMirValu
                                  XsMirError *error);
 XsMirStatus xs_mir_block_or_i32(XsMirBlock *block, XsMirValueId left, XsMirValueId right, XsMirValueId *result,
                                 XsMirError *error);
+XsMirStatus xs_mir_block_shl_i32(XsMirBlock *block, XsMirValueId left, XsMirValueId right, XsMirValueId *result,
+                                 XsMirError *error);
+XsMirStatus xs_mir_block_shr_i32(XsMirBlock *block, XsMirValueId left, XsMirValueId right, XsMirValueId *result,
+                                 XsMirError *error);
 XsMirStatus xs_mir_block_eq_i32(XsMirBlock *block, XsMirValueId left, XsMirValueId right, XsMirValueId *result,
                                 XsMirError *error);
 XsMirStatus xs_mir_block_ne_i32(XsMirBlock *block, XsMirValueId left, XsMirValueId right, XsMirValueId *result,