+++ xs/sources/mir/model_blocks.c
@@ -326,6 +326,20 @@ XsMirStatus xs_mir_block_or_i32(XsMirBlock *block, XsMirValueId left, XsMirValue
                         error);
 }
 
+XsMirStatus xs_mir_block_shl_i32(XsMirBlock *block, XsMirValueId left, XsMirValueId right, XsMirValueId *result,
+                                 XsMirError *error)
+{
+  return add_i32_binary(block, left, right, XS_MIR_INSTRUCTION_SHL_I32, (XsMirType){.kind = XS_LIL_TYPE_I32}, result,
+                        error);
+}
+
+XsMirStatus xs_mir_block_shr_i32(XsMirBlock *block, XsMirValueId left, XsMirValueId right, XsMirValueId *result,
+                                 XsMirError *error)
+{
+  return add_i32_binary(block, left, right, XS_MIR_INSTRUCTION_SHR_I32, (XsMirType){.kind = XS_LIL_TYPE_I32}, result,
+                        error);
+}
+
 XsMirStatus xs_mir_block_eq_i32(XsMirBlock *block, XsMirValueId left, XsMirValueId right, XsMirValueId *result,
                                 XsMirError *error)
 {