+++ xs/sources/xlil/model.c
@@ -496,6 +496,20 @@ XsLilStatus xs_lil_block_or_i32(XsLilBlock *block, XsLilValueId left, XsLilValue
                             (XsLilType){.kind = XS_LIL_TYPE_I32}, result, error);
 }
 
+XsLilStatus xs_lil_block_shl_i32(XsLilBlock *block, XsLilValueId left, XsLilValueId right, XsLilValueId *result,
+                                 XsLilError *error)
+{
+  return add_binary_integer(block, XS_LIL_INSTRUCTION_SHL_I32, left, right, XS_LIL_TYPE_I32,
+                            (XsLilType){.kind = XS_LIL_TYPE_I32}, result, error);
+}
+
+XsLilStatus xs_lil_block_shr_i32(XsLilBlock *block, XsLilValueId left, XsLilValueId right, XsLilValueId *result,
+                                 XsLilError *error)
+{
+  return add_binary_integer(block, XS_LIL_INSTRUCTION_SHR_I32, left, right, XS_LIL_TYPE_I32,
+                            (XsLilType){.kind = XS_LIL_TYPE_I32}, result, error);
+}
+
 XsLilStatus xs_lil_block_eq_i32(XsLilBlock *block, XsLilValueId left, XsLilValueId right, XsLilValueId *result,
                                 XsLilError *error)
 {