+++ xs/sources/xlil/writer.c
@@ -86,6 +86,14 @@ static XsLilStatus write_block(FILE *stream, XsLilError *error, const XsLilBlock
        fprintf(stream, "  %%r%u:i32 = or.i32 %%r%u, %%r%u\n", instruction->result, instruction->left,
                instruction->right) < 0)
       return xs_lil_set_error(error, XS_LIL_IO_ERROR, "could not write XLIL or.i32 instruction");
+    if(instruction->kind == XS_LIL_INSTRUCTION_SHL_I32 &&
+       fprintf(stream, "  %%r%u:i32 = shl.i32 %%r%u, %%r%u\n", instruction->result, instruction->left,
+               instruction->right) < 0)
+      return xs_lil_set_error(error, XS_LIL_IO_ERROR, "could not write XLIL shl.i32 instruction");
+    if(instruction->kind == XS_LIL_INSTRUCTION_SHR_I32 &&
+       fprintf(stream, "  %%r%u:i32 = shr.i32 %%r%u, %%r%u\n", instruction->result, instruction->left,
+               instruction->right) < 0)
+      return xs_lil_set_error(error, XS_LIL_IO_ERROR, "could not write XLIL shr.i32 instruction");
     if(instruction->kind == XS_LIL_INSTRUCTION_EQ_I32 &&
        fprintf(stream, "  %%r%u:bool = eq.i32 %%r%u, %%r%u\n", instruction->result, instruction->left,
                instruction->right) < 0)